Implemented Commandline Download Speedtest
Added the net/speedtest package that contains code for starting up a
speedtest server and a client. The speedtest command for starting the
server takes in a port number and maximum number of connections. It also
has a -localhost flag for if you want to use localhost rather than your
tailscale IP. The speedtest command for starting a client takes in
parameters for the speedtest as well as the host and port of the
speedtest server to connect to. Parameters include -size, which is the
message size, -time which is the test duration, -inc, which is the
interval length to measure the speed for (in seconds), as well as -u and
-d which are for upload and download test.

Signed-off-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>