Included random data in JSON encoded header
I was getting errors when trying to send the headers as JSON and the
randomly generated bytes seperately, so I included the random data in
the JSON message. I believe that this has a big effect on the reported
download speed.

This code is messy and filled with debug printlns and I am trying to
find a solution where the data is not JSON encoded.

Signed-off-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>